First Solo Data Project
The Amazon Stock prices dataset from 2012 to 2025 was analysed to inspect links and trends. A decision tree regressor was then created, trained, and tested achieving a Mean Squared Error of 1.34 and an R^2 Score of 0.9996, 99.96% (rounded up to 2 decimals) suggesting that the model is on average the difference between predicated and actual values is very small and that the model is .04% away from perfectly predicting the values, showing it is a highly reliable model.